refactor: add shared execute_hook helper
