+++ crates/hooktest/src/main.rs
@@ -1,3 +1,4 @@
+mod execute;
 mod notification;
 mod output;
 mod posttool;
